feat(router): add router and change structure
