+++ packages/client/index.html
@@ -2,12 +2,12 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/bomberman.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>{TODO: Set title :) }</title>
+    <title>Bomberman - Neverland edition</title>
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="./src/main.tsx"></script>
   </body>
 </html>
