--- packages/client/src/App.tsx
@@ -1,17 +1,0 @@
-import { useEffect } from 'react'
-import './App.css'
-
-function App() {
-  useEffect(() => {
-    const fetchServerData = async () => {
-      const response = await fetch('http://localhost:3001')
-      const data = await response.json()
-      console.log(data)
-    }
-
-    fetchServerData()
-  }, [])
-  return <div className="App">Вот тут будет жить ваше приложение :)</div>
-}
-
-export default App
