+++ packages/client/src/components/organisms/app/App.css
@@ -1,0 +1,19 @@
+.App {
+    background: linear-gradient(180deg, #595CAD 0%, #BC4154 100%);
+}
+
+.App:before {
+    content: ' ';
+    display: block;
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 1;
+    opacity: 0.03;
+    background-image: url('public/background.png');
+    background-repeat: no-repeat;
+    background-position: 50% 0;
+    background-size: cover;
+}
