+++ packages/client/src/components/pages/auth/Auth.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { Box } from '@mui/material';
+
+const Auth = () => {
+  return (
+    <Box
+    sx={{
+      height: '100vh',
+      display: 'flex',
+    }}
+    >
+        <p>Authorization page</p>
+    </Box>
+
+  );
+};
+
+export default Auth;
