+++ packages/client/src/components/pages/auth/index.ts
@@ -1,0 +1,1 @@
+export { default as AuthPage } from './Auth';
