+++ packages/client/src/index.css
@@ -1,0 +1,17 @@
+html,
+body,
+p {
+    margin: 0;
+    padding: 0;
+    overflow: hidden;
+}
+
+a {
+    text-decoration: none;
+
+    color: inherit;
+}
+
+* {
+    box-sizing: border-box;
+}
